requirements.txt
gspread>=6.2.0
google-auth>=2.40.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
playwright>=1.48.0
python-dotenv>=1.0.1
